+++ torchao/prototype/scaled_grouped_mm/benchmarks/benchmark_scaled_grouped_mm.py
